EZEE-2303: Moved InvalidXmlException to \eZ\Publish\Core\FieldType\RichText\Exception\
